Move appScheme into schemes folder
